requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Optional: Logging and Monitoring
langsmith>=0.1.0

# Optional: Evaluation and Analysis
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development Dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
